fix(web): share retained PR state with chat view
